backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-spectacular==0.28.0
psycopg2-binary
django-allauth==65.1.0
python-dotenv==1.0.1
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7